Do not create immer draft over undefined db
If an event was handled by the immer intereceptor before the store db was initialized immer would throw an error, since it's illegal to call createDraft on undefined. This was especially problematic when the immer interceptor was applied globally, since there was no easy way to opt out of it, which meant it was impossible to create the db in the first place.