Ensure all immer drafts get cleaned up
If a draft was created while processing an event but another interceptor or
event handler replaced db with something other than that draft, then
that draft would remain unfinished.

I'm not 100% sure that this would lead to a memory leak, since there
should be no references to the original draft, allowing it to be cleaned
up. However, I'm not 100% sure that's the case, so I'd prefer to be
cautious, and — regardless — it likely indicates a mistake by the user.